refactor: move og checker to backend
