+++ apps/dashboard/package.json
@@ -35,6 +35,7 @@
     "react-dom": "19.0.0-rc-66855b96-20241106",
     "satori": "^0.10.14",
     "sonner": "^1.4.41",
+    "use-debounce": "^10.0.4",
     "zundo": "^2.1.0",
     "zustand": "^4.5.2"
   },
